requirements.txt
Pillow==5.4.1
PyAudio==0.2.11
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
pyparsing==2.3.1
python-dateutil==2.8.0
scipy==1.2.1
setuptools==39.1.0
six==1.12.0